Se genera interfaz IResults llevando la logica de exportación a cada clase ExcelResults.cs y PDFResults.cs
