+++ Code/Source/App/Victoria.DesktopApp/Helpers/IResults.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Victoria.ViewModelWPF;
+
+namespace Victoria.DesktopApp.Helpers
+{
+    interface IResults
+    {
+        void Print();        
+    }
+}
