+++ Code/Source/App/Victoria.DesktopApp/View/AnalisisSensibilidadSimulacionPopUp.xaml.cs
@@ -24,9 +24,6 @@ using Victoria.Shared;
 using Victoria.Shared.AnalisisPrevio;
 using System.Collections.ObjectModel;
 using Victoria.DesktopApp.Helpers;
-//using excel = Microsoft.Office.Interop.Excel;
-using ClosedXML.Excel;
-using ClosedXML;
 
 namespace Victoria.DesktopApp.View
 {
@@ -129,12 +126,15 @@ namespace Victoria.DesktopApp.View
                         this.simulationStoped = true;
                         DateTime now = DateTime.Now;
                         this.simulationTotalTime = now.Subtract(this.simulationStartedTime);
-                        
-                        // Genero Resultados en formato Excel
-                        this.PrintResultsExcel(simulationPath + "\\resultados.xlsx");
 
-                        // GENERO LOS RESULTADOS EN PDF
-                        this.PrintResultsPDF(stages);
+                        //Genera lista con las diferentes exportaciones.
+                        List<IResults> lstExportarResultados = new List<IResults> {
+                                                                    new ExcelResults(this.simulationPath, "resultados.xlsx", this.stages, this.simulationTotalTime),
+                                                                    new PDFResults(this.simulationPath, "resultados.pdf", this.stages, this.simulationTotalTime)
+                                                                };
+
+                        //Se exportan los distintos formatos incluidos en la lista
+                        lstExportarResultados.ForEach(r => r.Print());                        
                         
                         // ABRO LA VENTANA DEL ARCHIVO QUE CONTIENE LOS RESULTADOS
                         System.Diagnostics.Process.Start("explorer.exe", "/select," + simulationPath);
@@ -161,236 +161,7 @@ namespace Victoria.DesktopApp.View
                     // asd
                     break;
             }
-        }
-
-        private void PrintResultsExcel(String fileName)
-        {
-            /*
-             * REQUERIMIENTO:
-             Lo que tiene que contener el archivo es una tabla donde las primeras columnas 
-             deben contener cada una de las variables de control y luego las variables de resultado. 
-             Cada una de las filas de esa tabla debe contener los valores para cada una de las 
-             corridas del análisis de sensibilidad. 
-             La primera fila (fila de títulos) debe contener el nombre de cada variable.
-            */
-            List<DataTable> resultsTable = createResultsTables(stages);
-
-            XLWorkbook oWB = new XLWorkbook();
-            var ws = oWB.Worksheets.Add("Resultados");            
-            object misvalue = System.Reflection.Missing.Value;            
-            int index_col = 1;
-            int index_row = 1;
-            int index_row_header = 4;
-            int cant_escenarios = 0;
-            bool tieneTituloVariables = false;
-
-            //Titulo
-            ws.Cell(index_row, index_col).Value = "Analisis de Sensibilidad";
-
-            index_row++;
-
-            ws.Row(2).Height = ws.Row(2).Height / 2;
-            /*
-             * La variable "resultsTable" contiene 2 tablas por cada escenario simulado, cada par de tablas
-             * contienen la siguiente SubTabla (en adelante tabla):
-             *      - La primer tabla contiene un TableName que representa el nombre del escenario y al 
-             *      recorrer las columnas y las filas podemos obtener información de las variables de control.
-             *      - La segunda tabla NO contiene un TableName, pero contiene datos de las variables de Resultado 
-             *      (que se pueden visualizar recorriendo las filas y columnas).
-             */
-            foreach (var tbl in resultsTable)
-            {                
-                if (tbl.TableName != "")
-                {
-                    //Obtengo información del escenario.
-                    cant_escenarios++;
-                    index_col = 1;
-                    ws.Cell(index_row_header + cant_escenarios, index_col).Value = tbl.TableName;                    
-                }
-                                                                
-                //Datos
-                for (int i = 0; i < tbl.Rows.Count; i++)
-                {
-                    index_col++;
-                    //Setea el nombre de las variables
-                    ws.Cell(index_row_header, index_col).Value = tbl.Rows[i][0].ToString();
-                    
-                    //En laprimer pasada setea el header (combinando celdas) indicando el tipo de variable
-                    if (!tieneTituloVariables)
-                    {
-                        ws.Cell(index_row_header-1 , index_col).Value = tbl.Rows[i].Table.Columns[0].ToString();
-                        ws.Range(index_row_header -1, index_col, index_row_header -1, index_col + tbl.Rows.Count - 1).Merge();
-                        ws.Cell(index_row_header - 1, index_col).Style.Alignment.WrapText = true;
-                        ws.Cell(index_row_header - 1, index_col).Style.Alignment.Vertical = XLAlignmentVerticalValues.Center;                        
-                        tieneTituloVariables = true;
-                    }
-                    //Setea el valor que tomaron las variables y lo coloca en filas diferentes en funcion del escenario.
-                    ws.Cell(index_row_header + cant_escenarios, index_col).Value = tbl.Rows[i][1].ToString();                    
-                }
-                tieneTituloVariables = false;
-            }
-
-            //Da formato al encabezado
-            ws.Range(1, 1, 1, index_col).Style.Fill.BackgroundColor = XLColor.Aquamarine;
-            ws.Range(1, 1, 1, index_col).Merge();
-            ws.Range(1, 1, 1, index_col).Style.Alignment.Horizontal = XLAlignmentHorizontalValues.Center;            
-            ws.Row(2).Style.Alignment.Horizontal = XLAlignmentHorizontalValues.Center;            
-            ws.Row(3).Style.Alignment.Horizontal = XLAlignmentHorizontalValues.Center;
-
-            //Se coloca el tiempo de ejecución
-            int index_tiempo_ejecucion = index_row_header + cant_escenarios + 2;
-            ws.Cell(index_tiempo_ejecucion, 1).Value = "Tiempo total de ejecución: " + this.simulationTotalTime.ToString(@"hh\:mm\:ss");
-            ws.Range(index_tiempo_ejecucion, 1, index_tiempo_ejecucion, index_col).Merge();
-
-            //Autoajustar al contenido
-            ws.Row(3).AdjustToContents(29.0,35.00);
-            ws.Columns().AdjustToContents(5.0,100.0);
-            
-            oWB.SaveAs(fileName);
-            
-        }       
-
-        private void PrintResultsPDF(IList<StageViewModelBase> stages)
-        {
-            List<DataTable> resultsTable = createResultsTables(stages);
-            
-            var filePath = simulationPath + "\\resultados.pdf";            
-            System.IO.FileStream fs = new System.IO.FileStream(filePath, FileMode.Create, FileAccess.Write, FileShare.None);
-            Document document = new Document();
-            document.SetPageSize(iTextSharp.text.PageSize.A4);
-            PdfWriter writer = PdfWriter.GetInstance(document, fs);
-            // calling PDFFooter class to Include in document
-            writer.PageEvent = new PDFFooter();
-
-            document.Open();
-
-            //Report Header
-            BaseFont bfntHead = BaseFont.CreateFont(BaseFont.TIMES_ROMAN, BaseFont.CP1252, BaseFont.NOT_EMBEDDED);
-            Font fntHead = new Font(bfntHead, 16, 1, BaseColor.BLACK);
-            iTextSharp.text.Paragraph prgHeading = new iTextSharp.text.Paragraph();
-            prgHeading.Alignment = Element.ALIGN_LEFT;
-            prgHeading.Add(new Chunk("Análisis de Sensibilidad", fntHead));
-            document.Add(prgHeading);
-
-            //Formato de textos
-            BaseFont bfntStage = BaseFont.CreateFont(BaseFont.TIMES_ROMAN, BaseFont.CP1252, BaseFont.NOT_EMBEDDED);
-            Font fntStage = new Font(bfntStage, 14, 1, BaseColor.BLACK);
-            BaseFont btnColumnHeader = BaseFont.CreateFont(BaseFont.TIMES_ROMAN, BaseFont.CP1252, BaseFont.NOT_EMBEDDED);
-            Font fntColumnHeader = new Font(btnColumnHeader, 14, 1, BaseColor.WHITE);
-
-            //CARGO LAS TABLAS SEGUN LA CANTIDAD DE STAGES
-            foreach (var tbl in resultsTable)
-            {
-                if (tbl.TableName != "")
-                {
-                    //LINEA
-                    iTextSharp.text.Paragraph p = new iTextSharp.text.Paragraph(new Chunk(new iTextSharp.text.pdf.draw.LineSeparator(0.0F, 100, BaseColor.BLACK, Element.ALIGN_LEFT, 0)));
-                    document.Add(p);
-                    document.Add(new Chunk("\n", fntHead));
-                    //NOMBRE DEL STAGE
-                    iTextSharp.text.Paragraph stageHeading = new iTextSharp.text.Paragraph();
-                    stageHeading.Alignment = Element.ALIGN_CENTER;
-                    stageHeading.Add(new Chunk(tbl.TableName, fntStage));
-                    document.Add(stageHeading);
-                }
-                document.Add(new Chunk("\n", fntStage));
-                //ESCRIBO TABLA
-                PdfPTable table = new iTextSharp.text.pdf.PdfPTable(tbl.Columns.Count);
-                //HEADER DE TABLA
-                for (int i = 0; i < tbl.Columns.Count; i++)
-                {
-                    PdfPCell cell = new iTextSharp.text.pdf.PdfPCell();
-                    cell.BackgroundColor = new BaseColor(67, 142, 185);
-                    cell.AddElement(new Chunk(tbl.Columns[i].ColumnName, fntColumnHeader));
-                    table.AddCell(cell);
-                }
-
-                //DATOS TABLA
-                for (int i = 0; i < tbl.Rows.Count; i++)
-                {
-                    for (int j = 0; j < tbl.Columns.Count; j++)
-                    {
-                        table.AddCell(tbl.Rows[i][j].ToString());
-                    }
-                }
-
-                document.Add(table);
-            }
-            //LINEA
-            iTextSharp.text.Paragraph line = new iTextSharp.text.Paragraph(new Chunk(new iTextSharp.text.pdf.draw.LineSeparator(0.0F, 100, BaseColor.BLACK, Element.ALIGN_LEFT, 0)));
-            document.Add(line);
-            document.Add(new Chunk("\n", fntHead));
-            //Tiempo total
-            BaseFont bfntTime = BaseFont.CreateFont(BaseFont.TIMES_ROMAN, BaseFont.CP1252, BaseFont.NOT_EMBEDDED);
-            Font fntTime = new Font(bfntTime, 14, 1, BaseColor.BLACK);
-            iTextSharp.text.Paragraph prgTime = new iTextSharp.text.Paragraph();
-            prgTime.Alignment = Element.ALIGN_LEFT;
-            prgTime.Add(new Chunk("Tiempo total de ejecución: "+this.simulationTotalTime.ToString(@"hh\:mm\:ss"), fntTime));
-            document.Add(prgTime);
-            
-            document.Close();
-            writer.Close();
-            fs.Close();
-        }
-
-        private List<DataTable> createResultsTables(IList<StageViewModelBase> stages)
-        {
-            List<DataTable> tablesList = new List<DataTable>();
-
-            foreach (var stg in stages)
-            {
-                DataTable table = new DataTable();
-                
-                table.TableName = stg.Name;
-                table.Columns.Add("Variable de Control");
-                table.Columns.Add("Valor");
-
-                DataTable table2 = new DataTable();
-
-                table2.TableName = "";
-                table2.Columns.Add("Variable de Resultado");
-                table2.Columns.Add("Valor");
-
-                foreach (var variable in stg.Simulation.GetVariables())
-                {
-                    if (variable.Type == VariableType.Control)
-                    {
-                        if (variable is VariableArray)
-                        {
-                            var variableArray = (VariableArray)variable;
-                            foreach (var variableAux in variableArray.Variables)
-                            {
-                                table.Rows.Add(variableAux.Name, variableAux.ActualValue.ToString());
-                            }
-                        }
-                        else
-                        {
-                            table.Rows.Add(variable.Name, variable.ActualValue.ToString());
-                        }
-                    }
-                    if (variable.Type == VariableType.Result)
-                    {
-                        if (variable is VariableArray)
-                        {
-                            var variableArray = (VariableArray)variable;
-                            foreach (var variableAux in variableArray.Variables)
-                            {
-                                table2.Rows.Add(variableAux.Name, variableAux.ActualValue.ToString());
-                            }
-                        }
-                        else
-                        {
-                            table2.Rows.Add(variable.Name, variable.ActualValue.ToString());
-                        }
-                    }
-                }
-
-                tablesList.Add(table);
-                tablesList.Add(table2);
-            }
-
-            return tablesList;
-        }
+        }        
 
         private void BtnCancel_OnClick(object sender, RoutedEventArgs e)
         {
